fix(*): few cleanups
